+++ include/roq/oms/order_update.hpp
@@ -13,6 +13,7 @@ struct ROQ_PUBLIC OrderUpdate final {
   std::string_view symbol;
   Side side = {};
   PositionEffect position_effect = {};
+  MarginMode margin_mode = {};
   double max_show_quantity = NaN;
   OrderType order_type = {};
   TimeInForce time_in_force = {};
@@ -56,6 +57,7 @@ struct fmt::formatter<roq::oms::OrderUpdate> {
         R"(symbol="{}", )"
         R"(side={}, )"
         R"(position_effect={}, )"
+        R"(margin_mode={}, )"
         R"(max_show_quantity={}, )"
         R"(order_type={}, )"
         R"(time_in_force={}, )"
@@ -87,6 +89,7 @@ struct fmt::formatter<roq::oms::OrderUpdate> {
         value.symbol,
         value.side,
         value.position_effect,
+        value.margin_mode,
         value.max_show_quantity,
         value.order_type,
         value.time_in_force,