+++ include/roq/order_update.hpp
@@ -15,6 +15,7 @@
 #include "roq/event.hpp"
 #include "roq/execution_instruction.hpp"
 #include "roq/liquidity.hpp"
+#include "roq/margin_mode.hpp"
 #include "roq/mask.hpp"
 #include "roq/name.hpp"
 #include "roq/numbers.hpp"
@@ -37,6 +38,7 @@ struct ROQ_PUBLIC OrderUpdate final {
   std::string_view symbol;                            //!< Symbol
   Side side = {};                                     //!< Side
   PositionEffect position_effect = {};                //!< Position effect
+  MarginMode margin_mode = {};                        //!< Margin mode
   double max_show_quantity = NaN;                     //!< Quantity visible to market (requires exchange support)
   OrderType order_type = {};                          //!< Order type
   TimeInForce time_in_force = {};                     //!< Time in force
@@ -91,6 +93,7 @@ struct fmt::formatter<roq::OrderUpdate> {
         R"(symbol="{}", )"
         R"(side={}, )"
         R"(position_effect={}, )"
+        R"(margin_mode={}, )"
         R"(max_show_quantity={}, )"
         R"(order_type={}, )"
         R"(time_in_force={}, )"
@@ -128,6 +131,7 @@ struct fmt::formatter<roq::OrderUpdate> {
         value.symbol,
         value.side,
         value.position_effect,
+        value.margin_mode,
         value.max_show_quantity,
         value.order_type,
         value.time_in_force,